requirements.txt
django==3.1.4
djangorestframework==3.12.2
django-currentuser==0.5.2
requests==2.25.1